==> src/notebooks/0_Launch_LS.ipynb <==
# Launch LABEL STUDIO

This notebook is used to launch Label Studio from a Jupyter server. To avoid conflicts if several users need to work at the same time, edit a new port for each of them. Label Studio's default port is 8080, so if you need to change it, you must check the YOURPORT variable and change it if necessary.
See the Label Studio documentation https://labelstud.io/guide/start.html#Run-Label-Studio-on-localhost-with-a-different-port

**Warning**
Before launching this notebook, make sure that the kernel you have chosen contains the necessary dependencies to launch Label Studio. If this is not the case, please follow the instructions below: 

**Notice concerning use** 
Any use, even partial, of the content of this notebook must be accompanied by an appropriate citation.

&copy; 2025 Marion Charpier

import os

# Default configuration for Label Studio
YOUR_PORT = "8080"
SERVER_IP = "localhost"

# Set Label Studio environment variables
os.environ['LABEL_STUDIO_PORT'] = YOUR_PORT
os.environ['LABEL_STUDIO_HOST'] = f'http://{SERVER_IP}:{YOUR_PORT}'
os.environ['LOCAL_FILES_DOCUMENT_ROOT'] = '/'
os.environ['LABEL_STUDIO_LOCAL_FILES_SERVING_ENABLED'] = 'true'

# Launch Label Studio
try:
    print("Starting Label Studio...")
    os.system('label-studio start')
except Exception as e:
    print(f"Error while starting Label Studio: {e}")
